Increase optin test timeout
